Introduce EDF saving

+++ src/EEGData.h
@@ -4,7 +4,7 @@
 #include "RC/Data1D.h"
 
 namespace CML {
-  typedef RC::Data1D<RC::Data1D<float>> EEGData;
+  typedef RC::Data1D<RC::Data1D<int16_t>> EEGData;
 }
 
 #endif // EEGDATA_H
